+++ src/main/java/com/mojang/minecraft/sound/Sound.java
@@ -16,7 +16,7 @@ public final class Sound implements Audio {
 	this.volume = var2.getDistanceSq() * var1.volume;
     }
 
-    public final boolean play(int[] var1, int[] var2, int var3) {
+    public final boolean play(short[] var1, short[] var2, int var3) {
 	if (data.length < var3) {
 	    data = new short[var3];
 	}
